update by lin t

+++ contracts/src/policy/PolicyGroup.sol
@@ -4,9 +4,10 @@ import "../market/MarketGroup.sol";
 import "../common/storage/UsingStorage.sol";
 
 contract PolicyGroup is UsingStorage {
-
 	function add(address _policy) external {
-		uint index = eternalStorage().getUint(keccak256("_policyGroupIndex"));
+		uint256 index = eternalStorage().getUint(
+			keccak256("_policyGroupIndex")
+		);
 		bytes32 key = getKey(index);
 		eternalStorage().setAddress(key, _policy);
 		index++;
@@ -17,21 +18,23 @@ contract PolicyGroup is UsingStorage {
 		return eternalStorage().getUint(keccak256("_policyGroupIndex"));
 	}
 
-	function get(uint _index) external view returns (address) {
+	function get(uint256 _index) external view returns (address) {
 		bytes32 key = getKey(_index);
 		return eternalStorage().getAddress(key);
 	}
 
 	function deleteAll() external {
-		uint index = eternalStorage().getUint(keccak256("_policyGroupIndex"));
-		for (uint i = 0; i < index; i++) {
+		uint256 index = eternalStorage().getUint(
+			keccak256("_policyGroupIndex")
+		);
+		for (uint256 i = 0; i < index; i++) {
 			bytes32 key = getKey(index);
 			eternalStorage().setAddress(key, address(0));
 		}
 		eternalStorage().setUint(keccak256("_policyGroupIndex"), 0);
 	}
 
-	function getKey(uint _index) private pure returns (bytes32) {
+	function getKey(uint256 _index) private pure returns (bytes32) {
 		return keccak256(abi.encodePacked("_policyGroup", _index));
 	}
 }
